chore: Update rules_python to 0.18.1
This is a recent version supporting Bazel 6 and the forward path for
managing external dependencies based on `pip_parse`. We don't use the
latest version as rules_python >= 0.19.0 is no longer compatible to
Bazel 4.x and it might still be too early for some users of this project
for this step.

Noteworthy changes:
- Use `pip_parse` instead of deprecated `pip_install`
- Require now a fully resolved requirements file
- There is no more default mypy shipped with this project. Users always
  have to provide a resolved requirements file for mypy.
- The mypy in third_party is used solely for testing